Added in some more functionality for generating training audio. However I have learned some new problems/questions
-I believe my first interpretation of how to use the synthSpeech and AddtoSynth functions was wrong. SynthSpeech will return the number of samples (size) of an audio file and the audio file. It accepts a number of frames and a framerate (where frames are states of the vocal tract)
AddtoSynth does not request this however, just the number of new samples to synthesize and the according parameters from the synthSpeech. My investigations showed that having mismatched counts of  parameter counts from SynthSpeech(tube areas) and user generated (tube lengths) did not have a noticeable effect, furthermore have more then 40(number tube sections) also did not noticeably affect the sound. MY Thoughts then, the model is just discarding excess values above the needed 40.
This suggests to me then that what we need to do is hand only a single frame to AddToSynth at a time. Which would still make sense for the description of the function.
So After experimenting with sending individual frames, the synthesiser still works, and respectufully well I think, however it does make me question what role the newglottis parameter has on the function now.

So after attempting the above, I found a bug that I do not yet know the cause of.
In attempts to assert if the resetSynthesis function worked.
-I used the human version, putting the reset synthesis inside the loop and outside, assuming the addToSynthesis would reset itself appropriately or keep building off itself.
My findings were:
With inside the loop, I was not reproducing the same sound between every loop,  but there would be some periodicity between 2-4 iterations.
Outside the loop, the periodicity was closer to 1, with the exception of the last audio out(i assumed i did something funny with writing the audio)
-Now for the bug. The results of the audio synthesis in both cases did not lend to reproduction... Neither case would come out exactly the same if the program was restarted. This also covers the case of if the reset is compeltely removed.
I am suspect there could be memory allocation issues (perhaps I am allocating something to too small a size, so I intend on going through the allocations and increasing size to see if the issue is rectified.)